Second release - added datasheet download functionality. Reformatted and commented code.
