--- README.md
@@ -1,2 +1,0 @@
-# IC-Inventory
-Python project to work with the Octopart API. Goal is to make a program that allows me to search for IC part numbers, and add them to a local inventory. Full details at learn-cnc.com
